requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pytesseract
pdf2image
Pillow
httpx
python-multipart
